requirements.txt
pdf2image==1.17.0
pillow==11.1.0
tqdm==4.67.1
pytest-timeout>=2.1.0
